#28930 fixing test

+++ dotcms-integration/src/test/java/com/dotmarketing/portlets/contentlet/business/ContentletAPITest.java
@@ -8825,6 +8825,6 @@ public class ContentletAPITest extends ContentletBaseTest {
         APILocator.getContentletAPI().setContentletProperty(publishedContentlet, new LegacyFieldTransformer(field).asOldField(), value);
 
         final Number valueRetrieved = (Number) publishedContentlet.get(keyPropName);
-        assertEquals(value, valueRetrieved);
+        assertEquals(value.longValue(), valueRetrieved);
     }
 }
